% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{acf_res}
\alias{acf_res}
\title{ACF on the residuals}
\usage{
acf_res(x, what)
}
\arguments{
\item{x}{Residual object as returned from one of the residual functions}

\item{what}{Dimension to use between "row", "column" or "diagonal"}
}
\description{
ACF on the residuals
}
\details{
Function used to estimate ACF for the 3 different dimensions of the residual matrix. 

It is assumed that the first composition group is in the first row and the last in the last row.
}
