% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{logistictransf}
\alias{logistictransf}
\title{Transform logistic-normal to normal observations}
\usage{
logistictransf(p, mult = FALSE)
}
\arguments{
\item{p}{Vector of mean values}

\item{mult}{Multiplicative logistic-normal if TRUE, otherwise additive}
}
\value{
Transformed observations
}
\description{
Transform logistic-normal to normal observations
}
\details{
This function transforms logistic-normal observations to normally distributed observations with dimension (length(p)-1).
}
