% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cres}
\alias{plot.cres}
\title{Plot residual object}
\usage{
\method{plot}{cres}(x, ...)
}
\arguments{
\item{x}{residual object as returned from one of the residual functions.}

\item{...}{extra arguments}
}
\description{
Plot residual object
}
\details{
gives a 4 plot overview of the residuals.
}
