% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Simulate Dirichlet observations}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{Number of simulations}

\item{alpha}{concentration parameter}
}
\value{
Matrix of Dirichlet observations
}
\description{
Simulate Dirichlet observations
}
\details{
The model ...
}
