% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rdirM}
\alias{rdirM}
\title{Simulate Dirichlet-multinomial observations}
\usage{
rdirM(n, N, alpha)
}
\arguments{
\item{n}{Number of simulations}

\item{N}{Sum of observations}

\item{alpha}{concentration parameter}
}
\value{
Matrix of Dirichlet-multinomial observations
}
\description{
Simulate Dirichlet-multinomial observations
}
\details{
The model ...
}
