% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{resDir}
\alias{resDir}
\title{Residuals for Dirichlet observations}
\usage{
resDir(obs, alpha, ...)
}
\arguments{
\item{obs}{Matrix of observations (proportions) assumed to be Dirichlet distributed}

\item{alpha}{Matrix of concentration parameters}

\item{...}{Additional arguments to pass to TMB::oneStepPredict}
}
\value{
One step ahead quantile residuals
}
\description{
Residuals for Dirichlet observations
}
\details{
The model estimates the one step ahead quantile residuals for observations that are Dirichlet distributed.
 
The matrices should be created such as each column is one sample from a Dirichlet distribution, the columns are assumed independent (i.e. the compositions are the rows).
}
\examples{
a<-matrix(rep(1000*c(.2,.2,.1,.1,.1,.3), 100), nrow=6)
o<-rdirichlet(100,a[,1])
res<-resDir(o,a)
plot(res)
}
