% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{resDirM}
\alias{resDirM}
\title{Residuals for Dirichlet-multinomial observations}
\usage{
resDirM(obs, alpha, ...)
}
\arguments{
\item{obs}{Matrix of observations assumed to be Dirichlet-multinomial distributed}

\item{alpha}{Matrix of concentration parameters}

\item{...}{Additional arguments to pass to TMB::oneStepPredict}
}
\value{
One step ahead randomized quantile residuals
}
\description{
Residuals for Dirichlet-multinomial observations
}
\details{
The model estimates the one step ahead randomized quantile residuals for observations that are Dirichlet-multinomial distributed.
 
The matrices should be created such as each column is one sample from a Dirichlet-multinomial distribution, the columns are assumed independent (i.e. the compositions are the rows).
}
\examples{
a<-matrix(rep(1000*c(.2,.2,.1,.1,.1,.3), 100), nrow=6)
o<-rdirM(100,1000, a[,1])
res<-resDirM(o,a)
plot(res)
}
