% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{reslogistN}
\alias{reslogistN}
\title{Residual Logistic-normal model}
\usage{
reslogistN(obs, mu, S, do_mult)
}
\arguments{
\item{obs}{An observation matrix (proportions) assumed to be Logistic-normal distributed}

\item{mu}{predicted observations}

\item{S}{covariance matrix, either one matrix to use for columns of obs or a list of covariance matrices for each column}

\item{do_mult}{0 if additive logistic-normal, 1 if multiplicative logistic-normal}
}
\value{
One step ahead randomized quantile residuals
}
\description{
Residual Logistic-normal model
}
\details{
The model ...
}
\examples{
mu <- matrix(0,ncol=100, nrow=5)
sigma <- diag(nrow(mu))
do_mult <- F
o <- sapply(1:100, function(x) rlogistN(mu[,1], sigma, do_mult))
res <- reslogistN(o, mu, sigma, do_mult)
plot(res)
}
