% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{resMulti}
\alias{resMulti}
\title{Residual multinomial model}
\usage{
resMulti(obs, pred)
}
\arguments{
\item{obs}{An observation matrix assumed to be multinomial}

\item{pred}{predicted observations}
}
\value{
One step ahead randomized quantile residuals
}
\description{
Residual multinomial model
}
\details{
The model ...
}
\examples{
o<-rmultinom(100, 25, c(.2,.2,.1,.1,.1,.3))
p<-matrix(rep(25*c(.2,.2,.1,.1,.1,.3), 100), nrow=6)
res<-resMulti(o,p)
plot(as.vector(res))
}
