% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{resmvnorm}
\alias{resmvnorm}
\title{Residual multivariate-normal model}
\usage{
resmvnorm(obs, mu, S, do_mult)
}
\arguments{
\item{obs}{An observation matrix assumed to be multivariate-normal distributed}

\item{mu}{predicted observations}

\item{S}{covariance matrix, either one matrix to use for columns of obs or a list of covariance matrices for each column}
}
\value{
One step ahead randomized quantile residuals
}
\description{
Residual multivariate-normal model
}
\details{
The model ...
}
\examples{
mu <- matrix(0,ncol=100, nrow=6)
sigma <- diag(nrow(mu))
o <- sapply(1:100, function(x) MASS::mvrnorm(1,mu[,k],sigma))
res <- resmvnorm(o, mu, sigma)
plot(res)
}
