% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rlogistN}
\alias{rlogistN}
\title{Simulate Logistic-normal observations}
\usage{
rlogistN(mu, sigma, mult = F)
}
\arguments{
\item{mu}{Vector of mean values}

\item{sigma}{Vector of standard errors}

\item{mult}{Multiplicative logistic-normal if TRUE, otherwise additive}
}
\value{
Matrix of logistic-normal observations
}
\description{
Simulate Logistic-normal observations
}
\details{
The model ...
}
