% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{resDir}
\alias{resDir}
\title{Residual Dirichlet model}
\usage{
resDir(obs, alpha)
}
\arguments{
\item{obs}{An observation matrix (proportions) assumed to be Dirichlet distributed}

\item{alpha}{concentration parameter}
}
\value{
One step ahead randomized quantile residuals
}
\description{
Residual Dirichlet model
}
\details{
The model ...
}
\examples{
a<-matrix(rep(1000*c(.2,.2,.1,.1,.1,.3), 100), nrow=6)
o<-rdirichlet(100,a[,1])
res<-resDir(o,a)
plot(res)
}
