% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{resDirM}
\alias{resDirM}
\title{Residual Dirichlet-multinomial model}
\usage{
resDirM(obs, alpha)
}
\arguments{
\item{obs}{An observation matrix (proportions) assumed to be Dirichlet-multinomial distributed}

\item{alpha}{concentration parameter}
}
\value{
One step ahead randomized quantile residuals
}
\description{
Residual Dirichlet-multinomial model
}
\details{
The model ...
}
\examples{
a<-matrix(rep(1000*c(.2,.2,.1,.1,.1,.3), 100), nrow=6)
o<-rdirM(100,1000, a[,1])
res<-resDirM(o,a)
plot(res)
}
