% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{resMulti}
\alias{resMulti}
\title{Residual multinomial model}
\usage{
resMulti(obs, pred)
}
\arguments{
\item{obs}{An observation assumed to be multinomial}

\item{pred}{predicted observations}
}
\value{
One step ahead randomized quantile residuals
}
\description{
Residual multinomial model
}
\details{
The model ...
}
