% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{resmvnorm}
\alias{resmvnorm}
\title{Residuals for multivariate-normal observations}
\usage{
resmvnorm(obs, mu, S, ...)
}
\arguments{
\item{obs}{Matrix of observations assumed to be multivariate-normal distributed}

\item{mu}{Matrix of predicted observations}

\item{S}{Covariance matrix, either one matrix to use for columns of obs or a list of covariance matrices for each column}

\item{...}{Additional arguments to pass to TMB::oneStepPredict}
}
\value{
One step ahead quantile residuals
}
\description{
Residuals for multivariate-normal observations
}
\details{
The model estimates the one step ahead quantile residuals for observations that are multivariate-normal distributed.
 
The matrices should be created such as each column is one sample from a multivariate-normal distribution, the columns are assumed independent (i.e. the compositions are the rows).
}
\examples{
mu <- matrix(0,ncol=100, nrow=6)
sigma <- diag(nrow(mu))+0.5
o <- replicate(100, MASS::mvrnorm(1,mu[,1],sigma))
res <- resmvnorm(o, mu, sigma)
plot(res)
}
