% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rlogistN}
\alias{rlogistN}
\title{Simulate logistic-normal observations}
\usage{
rlogistN(mu, Sigma, mult = FALSE)
}
\arguments{
\item{mu}{Vector of mean values}

\item{Sigma}{Covariance matrix}

\item{mult}{Multiplicative logistic-normal if TRUE, otherwise additive}
}
\value{
Matrix of logistic-normal observations
}
\description{
Simulate logistic-normal observations
}
\details{
This function simulates logistic-normal observations.
}
